Move the current pipeline to a sequential module
