+++ components/sic_cli/src/pipeline/mod.rs
@@ -1,0 +1,2 @@
+/// A sequential pipeline
+pub mod seq;
